@@ -1,10 +1,10 @@
 name: bienaventurados
-description: A new Flutter project.
+description: Aplicación católica.
 publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   cloud_firestore: ^2.5.3
@@ -30,17 +30,34 @@ dependencies:
 dev_dependencies:
   build_runner: null
   flutter_launcher_icons: ^0.9.0
+  flutter_native_splash: ^1.2.4
   flutter_test:
     sdk: flutter
   hive_generator: null
+
+flutter_native_splash:
+  android: true
+  # ios: true
+  # web: true
+
+  color: "#F9FAFE"
+  image: "assets/splash.png"
+
+  android_gravity: center
+  # ios_content_mode: center
+  # web_image_mode: center
+
+  #color_dark: "#232733"
+  #image_dark: "assets/images/dark_splash.png"
+  fullscreen: true
 
 flutter_icons:
   image_path_android: "assets/images/ic_launcher.png"
   image_path_ios: "assets/images/ic_launcher.png"
   android: true
   ios: true
-  adaptive_icon_background: "#9147fe"
-  adaptive_icon_foreground: "assets/images/ic_foreground_variant.png"
+  adaptive_icon_background: "#F9FAFE"
+  adaptive_icon_foreground: "assets/images/ic_foreground.png"
 
 flutter:
   uses-material-design: true
@@ -49,16 +66,6 @@ flutter:
     - assets/images/
 
   fonts:
-    - family: Montserrat
-      fonts:
-        - asset: assets/fonts/Montserrat-Regular.ttf
-        - asset: assets/fonts/Montserrat-Italic.ttf
-          style: italic
-        - asset: assets/fonts/Montserrat-Medium.ttf
-          weight: 500
-        - asset: assets/fonts/Montserrat-Bold.ttf
-          weight: 700
-
     - family: Gotham
       fonts:
         - asset: assets/fonts/Gotham-Book.ttf
