@@ -167,7 +167,7 @@ class _RegistrarsePageState extends State<RegistrarsePage> {
                     style: Theme.of(context)
                         .textTheme
                         .headline1!
-                        .copyWith(color: Theme.of(context).accentColor),
+                        .copyWith(color: Theme.of(context).colorScheme.secondary),
                   ),
                 ),
               ],
