@@ -26,7 +26,7 @@ class ThemeProvider extends ChangeNotifier{
       // overline: TextStyle(decoration: TextDecoration.underline, fontFamily: 'Gotham', fontSize: 12, fontWeight: FontWeight.normal, fontStyle: FontStyle.normal, color: Colores.primarioNight),
     ),
     fontFamily: 'Gotham',
-    accentColor: Colores.acento,
+    //accentColor: Colores.acento,
     textSelectionTheme: TextSelectionThemeData(
       cursorColor: Colores.acento
     ),
@@ -38,12 +38,27 @@ class ThemeProvider extends ChangeNotifier{
         padding: MaterialStateProperty.all<EdgeInsetsGeometry>(EdgeInsets.symmetric(horizontal: 20)),
       )
     ),
+    colorScheme: ColorScheme(
+      primary: Colores.primarioDay,
+      primaryVariant: Colores.primarioNight,
+      secondary: Colores.acento,
+      surface: Colores.primarioDay,
+      background: Colors.deepOrange,
+      error: Colors.greenAccent,
+      secondaryVariant: Colors.yellowAccent,
+      brightness: Brightness.light,
+      onBackground: Colors.red,
+      onError: Colors.blue,
+      onPrimary: Colores.primarioNight,
+      onSecondary: Colors.green,
+      onSurface: Colores.primarioNight,
+    ),
     scaffoldBackgroundColor: Colores.primarioDay,
     primaryColor: Colores.primarioDay,
-    disabledColor: Colores.secundarioDay,
-    cardColor: Colores.secundarioDay,
-    brightness: Brightness.light,
-    dividerColor: Colores.contrasteDay,
+    // disabledColor: Colores.secundarioDay,
+    // cardColor: Colores.secundarioDay,
+    // brightness: Brightness.light,
+    // dividerColor: Colores.contrasteDay,
     hoverColor: Colores.contrasteDay, // color de contraste
     highlightColor: Colors.transparent, // resaltados y tabs
     dialogBackgroundColor: Colores.primarioDay,
@@ -71,7 +86,7 @@ class ThemeProvider extends ChangeNotifier{
       // overline: TextStyle(decoration: TextDecoration.underline, fontFamily: 'Gotham', fontSize: 12, fontWeight: FontWeight.normal, fontStyle: FontStyle.normal, color: Colores.primarioNight),
     ),
     fontFamily: 'Montserrat',
-    accentColor: Colores.acento,
+    //accentColor: Colores.acento,
     textSelectionTheme: TextSelectionThemeData(
       cursorColor: Colores.acento 
     ),
@@ -81,15 +96,30 @@ class ThemeProvider extends ChangeNotifier{
         shape: MaterialStateProperty.all<OutlinedBorder>(CircleBorder()),
       )
     ),
+    colorScheme: ColorScheme(
+      primary: Colores.primarioNight,
+      primaryVariant: Colores.primarioDay,
+      secondary: Colores.acento,
+      surface: Colores.primarioNight,
+      background: Colors.deepOrange,
+      error: Colors.greenAccent,
+      secondaryVariant: Colores.primarioDay,
+      brightness: Brightness.dark,
+      onBackground: Colors.red,
+      onError: Colors.blue,
+      onPrimary: Colores.primarioDay,
+      onSecondary: Colors.green,
+      onSurface: Colores.primarioDay,
+    ),
     scaffoldBackgroundColor: Colores.primarioNight,
     primaryColor: Colores.primarioNight,
-    disabledColor: Colores.secundarioNight,
-    cardColor: Colores.secundarioNight,
-    brightness: Brightness.dark,
-    dividerColor: Colores.contrasteNight,
+    //disabledColor: Colores.secundarioNight,
+    //cardColor: Colores.secundarioNight,
+    //brightness: Brightness.dark,
+    //dividerColor: Colores.contrasteNight,
     hoverColor: Colores.contrasteNight,
-    hintColor: Colores.contrasteNight,
-    shadowColor: Colores.secundarioNight,
+    //hintColor: Colores.contrasteNight,
+    //shadowColor: Colores.secundarioNight,
     highlightColor: Colors.transparent,
     dialogBackgroundColor: Colores.primarioNight,
     primaryColorDark: Colores.contrasteNight,
